% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envKernel.R
\name{EnvKernel}
\alias{EnvKernel}
\title{Easily Building of Environmental Relatedness Kernels}
\usage{
EnvKernel(
  df.cov,
  Y = NULL,
  is.scaled = T,
  sd.tol = 1,
  tol = 0.001,
  bydiag = FALSE,
  merge = FALSE,
  env.id = NULL
)
}
\arguments{
\item{df.cov}{matrix. Data from environmental variables (or markers) per environment (or combinations of genotype-environment)}

\item{is.scaled}{boolean. If environmental data is mean-centered and scaled (default = TRUE), assuming x~N(0,1)}

\item{sd.tol}{numeric. Maximum standard deviation value for quality control. Coluns above this value are eliminated}
}
\description{
Returns environmental kinships for reaction norm models. Output is a list containing the objects varCov kinship for environmental variables and envCov kinshp for environmental relatedness
}
\author{
Germano Costa Neto
}
