% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covfromraster.R
\name{Extract_GIS}
\alias{Extract_GIS}
\title{Easily Extraction of Worldwide Raster-based Data}
\usage{
Extract_GIS(covraster = NULL, env.data = NULL)
}
\arguments{
\item{covraster}{RasterLayer. A raster from which data is going to be extracted.}

\item{env.data}{data.frame. A \code{get_weather} output.}
}
\value{
The original dataframe with additional geographic/weather information.
}
\description{
A wrapper of \code{raster::getData()} for extracting variables from rasters based on latitude and loongitude of experiments.
}
\details{
TODO
}
\examples{
# TODO

}
\author{
Germano Costa Neto
}
