% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGEenriched.R
\name{get_kernel}
\alias{get_kernel}
\title{Envirotype-informed kernels for statistical models}
\usage{
get_kernel(
  K_E = NULL,
  K_G = NULL,
  data = NULL,
  model = NULL,
  intercept.random = FALSE,
  reaction = FALSE,
  dimension_KE = NULL,
  ne = NULL,
  ng = NULL,
  env = "env",
  gid = "gid",
  y = "value"
)
}
\arguments{
\item{K_E}{list. Contains n matrices of envirotype-related kernels (n x n genotypes-environment). If NULL, benchmarck genomic kernels are built.}

\item{K_G}{list. Contains matrices of genomic enabled kernels (p x p genotypes). See BGGE::getK for more information.}

\item{data}{data.frame. Should contain the following colunms: environemnt, genotype, phenotype.}

\item{model}{character. Model structure for genomic predicion. It can be \code{c('MM','MDs','EMM','EMDs')}, in which MM (main effect model \eqn{y = m + Z_{E} + G}) and MDs (\eqn{y = m + Z_{E} + G + GE}).}

\item{intercept.random}{boolean. Indicates the inclusion of a genomic random intercept (default = FALSE). For more details, see BGGE package vignette.}

\item{reaction}{boolean. Indicates the inclusion of a reaction norm based GxE kernel (default = FALSE).}

\item{dimension_KE}{character. \code{size_E=c('q','n')}. In the first, 'q' means taht the environmental relationship kernel has the dimensions of q x q observations,in which q is the number of environments. If 'n', the relationship kernel has the dimension n=pq, in which p is the number of genotypes}

\item{ne}{numeric. denotes the number of environments (q)}

\item{ng}{numeric. denotes the number of genotypes (p)}

\item{env}{character. denotes the name of the column respectively to environments}

\item{gid}{character. denotes the name of the column respectively to genotypes}

\item{y}{character. denotes the name of the column respectively to phenotype values}
}
\value{
A list of kernels (relationship matrices) to be used in genomic models.
}
\description{
Get multiple genomic and/or envirotype-informed kernels for bayesian genomic prediciton.
}
\details{
Detail on the model structures is given in the GitHub page: {https://github.com/allogamous/EnvRtype/blob/master/Prediction.md}
}
\examples{
\dontrun{
### Loading the genomic, phenotype and weather data
data('maizeG'); data('maizeYield'); data("maizeWTH")

data("maizeYield") # toy set of phenotype data (grain yield per environment)
data("maizeG"    ) # toy set of genomic relationship for additive effects
data("maizeWTH")   # toy set of environmental data
y   = "value"      # name of the vector of phenotypes
gid = "gid"        # name of the vector of genotypes
env = "env"        # name of the vector of environments
ECs  = W_matrix(env.data = maizeWTH[maizeWTH$daysFromStart < 100, ], var.id = c("FRUE",'PETP',"SRAD","T2M_MAX"),statistic = 'mean')
## KG and KE might be a list of kernels
KE = list(W = env_kernel(env.data = ECs)[[2]])
KG = list(G=maizeG);
## Creating kernel models with get_kernel
## y = fixed + Genomic + error (main effect model, MM)
MM    = get_kernel(K_G = KG, y = y,gid = gid,env = env, data = maizeYield,model = "MM")
## y = fixed + Genomic + Genomic x Environment + error (MM plus a single GE deviation, MDs model, assuming GE as a block diagonal genomic effects)
MDs   = get_kernel(K_G = KG, y = y,gid = gid,env = env,  data = maizeYield, model = "MDs")
EMM   = get_kernel(K_G = KG, K_E = KE, y = y,gid = gid,env = env,  data = maizeYield, model = "EMM")
EMDs  = get_kernel(K_G = KG, K_E = KE, y = y,gid = gid,env = env,  data = maizeYield, model = "EMDs")
RMMM  = get_kernel(K_G = KG, K_E = KE, y = y,gid = gid,env = env,  data = maizeYield, model = "RNMM")
RNMDs = get_kernel(K_G = KG, K_E = KE, y = y,gid = gid,env = env,  data = maizeYield, model = "RNMDs")
## Examples of Models without any genetic relatedness, which G = I a identity for genotypes
MDs   = get_kernel(K_G = NULL, y = y,gid = gid,env = env,  data = maizeYield, model = "MDs")
EMDs  = get_kernel(K_G = NULL, K_E = KE, y = y,gid = gid,env = env,  data = maizeYield, model = "EMDs")
}

}
\seealso{
env_typing W_matrix kernel_model
}
\author{
Germano Costa Neto
}
