% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_soil.R
\name{get_soil}
\alias{get_soil}
\title{Easily Collection of Soil Features (250m res)}
\usage{
get_soil(
  env.id,
  lat,
  lon,
  home.path = NULL,
  variables.names = "clay",
  sleep = 20,
  logfile = NULL
)
}
\arguments{
\item{env.id}{vector (character or level). Identification of the site/environment (e.g. Piracicaba01).}

\item{lat}{vector (numeric). Latitude values of the site/environment (e.g. -13.05) in WGS84.}

\item{lon}{vector (numeric). Longitude values site/environment (e.g. -56.05) in WGS84.}

\item{variables.names}{vector (character).}

\item{properties}{A SoilGrid property. Defaults to \code{clay}. Possible}

\item{verbose}{Report messages to console? Defaults to \code{FALSE}.}
}
\value{
A `data.frame` object
}
\description{
Imports soil features from SoilGrids data set
}
\examples{
\dontrun{
## Downloading clay and nitrogen for a given location
get_soil(env.id = "NM", lat = -13.05, lon = -56.05,variables.names = c('clay','nitrogen'))

## download all variables (default) for a single location
get_soil(env.id = "NM", lat = -13.05, lon = -56.05)

## All variables for two locations
env = c("NM","SO")
lat = c(-13.05,-12.32); lon = c(-56.05,-55.42)
get_soil(env.id = env, lat = lat, lon = lon)
}

}
\references{
Poggio et al (2021). _SoilGrids 2.0: producing soil information for the globe with quantified spatial uncertainty_. European Geosciences Union, <https://doi.org/10.5194/soil-7-217-2021>.
}
\author{
Brandon Monier, modified by Germano Costa Neto
}
