% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_weather_gis.R
\name{get_weather}
\alias{get_weather}
\title{Easily Collection of Worldwide Daily Weather Data}
\usage{
get_weather(
  env.id = NULL,
  lat = NULL,
  lon = NULL,
  start.day = NULL,
  end.day = NULL,
  variables.names = NULL,
  dir.path = NULL,
  save = FALSE,
  asdataframe = FALSE,
  temporal.scale = "DAILY"
)
}
\arguments{
\item{env.id}{vector (character or level). Indicates site/environment identification}

\item{lat}{vector (numeric). Indicates de latitude valuesort}

\item{lon}{vector (numeric). containing longitude values}

\item{start.day}{start point}

\item{end.day}{end point}

\item{variables.names}{list of variables names}

\item{dir.path}{output directorie}

\item{save}{bollean. If TRUE, save each environmental data.frame as .csv in dir.path}

\item{temporal.scale}{character. Default = 'DAILY'. See get_power() function in nasapower package for more details}
}
\description{
Imports daily-scale weather data from the NASA POWER GIS database
}
\author{
Germano Costa Neto
}
