% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maizeWTH.R
\docType{data}
\name{maizeWTH}
\alias{maizeWTH}
\title{Maize Weather Data}
\format{
A data.frame containing weather variables for each environment in HEL dataset
}
\usage{
data(maizeWTH)
}
\description{
Weather data collected from NASA POWER for each one of the five environments in HELIX dataset.
}
\examples{
data(maizeWTH); head(maizeWTH)

}
\keyword{datasets}
