% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SradPARAM.R
\name{param_radiation}
\alias{param_radiation}
\title{Support functions to estimate basic radiation parameters}
\usage{
param_radiation(
  env.data,
  day.id = "DOY",
  latitude = "LAT",
  solarRadiation = NULL,
  merge = TRUE
)
}
\arguments{
\item{env.data}{data.frame. A \code{get_weather()} output or A \code{get_weather()}-like dataframe.}

\item{merge}{boolean. If \code{TRUE}, calculated variables are merged to the original dataframe. Default is TRUE}
}
\value{
Returns a dataframe with parameters related to solar radiation. See details for further information.
}
\description{
Core of functions to estimate atmospheric parameters related to solar radiation. Fore more details about the equations see de FAO-evapotranspiration publication http://www.fao.org/3/X0490E/x0490e07.htm#radiation.
}
\details{
This functions requires day of the year, latitude, thermal radiative flux, and sky insolation which are provided by \code{get_weather()}. If one of these parameters are missing, an error will be returned.
\itemize{
 \item n: Actual duration of sunshine (hour)
 \item N: Daylight hours (hour)
 \item RTA: Extraterrestrial radiation (MJ/m^2/day)
 \item SRAD: Solar radiation (MJ/m^2/day)
 }
}
\examples{
\dontrun{
### Fetching weather information from NASA-POWER
env.data = get_weather(lat = -13.05, lon = -56.05, country = 'BRA')

### Calculating solar radiation
param_radiation(env.data)

env.data = get_weather(lat = -13.05, lon = -56.05, country = 'BRA') \%>\%
           param_radiation()

#' env.data = get_weather(lat = -13.05, lon = -56.05, country = 'BRA') \%>\%
           param_radiation(day.id='DOY',latitude='LAT')
}
}
\author{
Germano Costa Neto
}
