% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/env_prediction.R
\name{print.kernel_set}
\alias{print.kernel_set}
\title{Print a kernel set}
\usage{
\method{print}{kernel_set}(x, ...)
}
\arguments{
\item{x}{object of class \code{"kernel_set"} from \code{\link{get_kernel}}.}

\item{...}{ignored.}
}
\description{
Print a kernel set
}
