% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/env_prediction.R
\name{print.sim_envirome}
\alias{print.sim_envirome}
\title{Print a Simulated Envirome Diversity Matrix}
\usage{
\method{print}{sim_envirome}(x, ...)
}
\arguments{
\item{x}{a \code{sim_envirome} object.}

\item{...}{ignored.}
}
\value{
\code{x}, invisibly.
}
\description{
Print a Simulated Envirome Diversity Matrix
}
\seealso{
\code{\link{sim_envirome}}
}
