% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/env_prediction.R
\name{print.sim_markers}
\alias{print.sim_markers}
\title{Print a Simulated Marker Matrix}
\usage{
\method{print}{sim_markers}(x, ...)
}
\arguments{
\item{x}{a \code{sim_markers} object.}

\item{...}{ignored.}
}
\value{
\code{x}, invisibly.
}
\description{
Print a Simulated Marker Matrix
}
\seealso{
\code{\link{sim_markers}}
}
