% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/env_prediction.R
\name{print.sim_met}
\alias{print.sim_met}
\title{Print a Simulated Multi-Environment Trial}
\usage{
\method{print}{sim_met}(x, ...)
}
\arguments{
\item{x}{a \code{sim_met} object.}

\item{...}{ignored.}
}
\value{
\code{x}, invisibly.
}
\description{
Print a Simulated Multi-Environment Trial
}
\seealso{
\code{\link{sim_met}}
}
