% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/env_prediction.R
\name{print.sim_W}
\alias{print.sim_W}
\title{Print a Simulated Envirome}
\usage{
\method{print}{sim_W}(x, ...)
}
\arguments{
\item{x}{a \code{sim_W} object.}

\item{...}{ignored.}
}
\value{
\code{x}, invisibly.
}
\description{
Print a Simulated Envirome
}
\seealso{
\code{\link{sim_W}}
}
