% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Wmatrix.R
\name{W.matrix}
\alias{W.matrix}
\title{Build the Matrix of Environmental Covariables for Reaction Norm}
\usage{
W.matrix(
  weather.data,
  is.processed = FALSE,
  id.names = NULL,
  env.id = NULL,
  var.id = NULL,
  probs = NULL,
  by.interval = NULL,
  time.window = NULL,
  names.window = NULL,
  center = T,
  scale = T,
  sd.tol = 10,
  statistic = NULL,
  tol = 0.001,
  QC = FALSE
)
}
\arguments{
\item{weather.data}{data.frame of environmental variables gotten from get_weather.}

\item{id.names}{vector (character). Indicates the name of the columns to be used as id for the environmental variables to be analysed.}

\item{env.id}{vector (character). Indicates the name of the columns to be used as id for environments.}

\item{var.id}{character. Indicates which variables will be used in the analysis.}

\item{probs}{vector(numeric). Indicates the probability quantiles, as probs = {0,1}, if cardinals is NULL. If is quantiles=NULL, quantiles = c(0.01,.25,.50,.99).}

\item{by.interval}{boolean. Indicates if temporal intervals must be computed insied of each environment. Default = FALSE.}

\item{time.window}{vector (numeric). If by.interval = TRUE, this argument indicates the temporal breaks for delimited intervals.}

\item{names.window}{vector(character). If by.interval = TRUE, this argument indicates the names of the desirable intervals.}

\item{scale}{boolean. If scale=TRUE, the variables (x) assumes a mean-centered scaled distribution, with x~N(0,1).}

\item{sd.tol}{numeric. Default value equal to 10.}

\item{statistic}{vector (character). Indicates what statistic must be runned, statistic = c('all','sum','mean','quantile'). Default: 'mean'.}

\item{QT}{boolean. Indicates with Quality Control is applied. QC is based on the standard deviation tolerance (sd.tol), removing variables (x) with sd(x) > sd.tol}
}
\description{
Returns a covariable realzed matrix with dimensions of q x k, for q environments and k covariables.
}
\author{
Germano Costa Neto
}
