% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clay_content.R
\docType{data}
\name{clay_5_15}
\alias{clay_5_15}
\title{Clay content (g/kg) from 5 to 15cm depth for Nairobi, Kenya}
\format{
A .tif raster file
}
\usage{
data(clay_5_15)
}
\description{
Soil profile for clay content (g/kg) for Nairobi using a raster file was downloaded from SoilGrids (https://soilgrids.org/).
}
\examples{
data(clay_5_15)
require(raster)
plot(clay_5_15)

}
\keyword{datasets}
