% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covfromraster.R
\name{Extract_GIS}
\alias{Extract_GIS}
\title{Easily Extraction of Worldwide Raster-based Data}
\usage{
Extract_GIS(covraster = NULL, weather.data = NULL)
}
\arguments{
\item{covraster}{RasterLayer. A raster from which data is going to be extracted.}

\item{weather.data}{data.frame. A get_weather output}

\item{K_E}{list of envirotype-related kernels (n x n genotypes-environment).
If NULL, benchmarck genomic kernels are built.}

\item{K_G}{list of genomic enabled kernels (p x p genotypes)}

\item{Y}{data.frame contaning the following colunms: environemnt, genotype, trait value}

\item{model}{model structure for genomic predicion. It can be c('MM','MDs','E-MM','E-MDs'),
which MM (main effect model or Y=fixed + G) amd MDs (Y=fixed+G+GxE)}

\item{reaction}{boolean, inclusion of a reaction norm based GxE kernel (default = FALSE)}

\item{intercept.random}{boolean, inclusion of a genomic random intercepet (default = FALSE)}
}
\description{
Extracts variables from rasters.
}
\author{
Germano Costa Neto
}
