% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covfromraster.R
\name{extract_GIS}
\alias{extract_GIS}
\title{Easily Extraction of Worldwide Raster-based Data}
\usage{
extract_GIS(
  covraster = NULL,
  Latitude = NULL,
  Longitude = NULL,
  env.data = NULL,
  env.id = NULL,
  name.out = NULL
)
}
\arguments{
\item{covraster}{RasterLayer. A raster from which data is going to be extracted.}

\item{Latitude}{character, the name of latitude column in env.data}

\item{Longitude}{character, the name of longitude column in env.data}

\item{env.data}{data.frame. A \code{get_weather} output.}

\item{env.id}{character, the name of environmental identification column in env.data}

\item{name.out}{character, output name}
}
\value{
The original dataframe with additional geographic/weather information.
}
\description{
A wrapper of \code{raster::getData()} for extracting variables from rasters based on latitude and loongitude of experiments.
}
\details{
TODO
}
\examples{
# TODO
# Extraction of clay soil content for Nairobi, Kenya
require(EnvRtype)
data("clay_5_15")
env.data = data.frame(LAT = -1.367, LON = 36.834, env = 'NAIROBI')
env.data = extract_GIS(covraster = clay_5_15,name.out = 'clay_5_15',env.data = env.data)
head(env.data)

}
\author{
Germano Costa Neto
}
