% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covfromraster.R
\name{Extract_GIS}
\alias{Extract_GIS}
\title{Easily Extraction of Worldwide Raster-based Data}
\usage{
Extract_GIS(
  covraster = NULL,
  reference = NULL,
  lon = NULL,
  lat = NULL,
  env.id = NULL,
  .crs = NULL,
  .path = NULL,
  covname = NULL
)
}
\arguments{
\item{K_E}{list of envirotype-related kernels (n x n genotypes-environment).
If NULL, benchmarck genomic kernels are built.}

\item{K_G}{list of genomic enabled kernels (p x p genotypes)}

\item{Y}{data.frame contaning the following colunms: environemnt, genotype, trait value}

\item{model}{model structure for genomic predicion. It can be c('MM','MDs','E-MM','E-MDs'),
which MM (main effect model or Y=fixed + G) amd MDs (Y=fixed+G+GxE)}

\item{reaction}{boolean, inclusion of a reaction norm based GxE kernel (default = FALSE)}

\item{intercept.random}{boolean, inclusion of a genomic random intercepet (default = FALSE)}
}
\description{
Get multuple genomic and envirotype-informed kernels for bayesian genomic prediciton
}
\author{
Germano Costa Neto
}
