% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGEenriched.R
\name{get_kernel}
\alias{get_kernel}
\title{Envirotype-informed kernels for statistical models}
\usage{
get_kernel(
  K_E = NULL,
  K_G,
  Y,
  model = NULL,
  intercept.random = FALSE,
  reaction = FALSE,
  size_E = NULL
)
}
\arguments{
\item{K_E}{list of envirotype-related kernels (n x n genotypes-environment).
If NULL, benchmarck genomic kernels are built.}

\item{K_G}{list of genomic enabled kernels (p x p genotypes)}

\item{Y}{data.frame contaning the following colunms: environemnt, genotype, trait value}

\item{model}{model structure for genomic predicion. It can be c('MM','MDs','E-MM','E-MDs'),
which MM (main effect model or Y=fixed + G) amd MDs (Y=fixed+G+GxE)}

\item{intercept.random}{boolean, inclusion of a genomic random intercept (default = FALSE). For more details, see BGGE package vignette.}

\item{reaction}{boolean, inclusion of a reaction norm based GxE kernel (default = FALSE)}

\item{size_E}{character. size_E=c('full','environment'). In the first, 'full' means taht the environmental relationship kernel has the dimensions of n x n observations, which n = pq (p genotypes, q environments). If 'environment' the size of E-kernel is q x q.}
}
\description{
Get multiple genomic and/or envirotype-informed kernels for bayesian genomic prediciton
}
\author{
Germano Costa Neto
}
