% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/met_kernel_model.R
\name{kernel_model}
\alias{kernel_model}
\title{Kernel Models for Predicting Phenotypes across Multi-Environment Conditions}
\usage{
kernel_model(
  y,
  data = NULL,
  random = NULL,
  fixed = NULL,
  env,
  gid,
  verbose = FALSE,
  iterations = 1000,
  burnin = 200,
  thining = 10,
  tol = 1e-20,
  R2 = 0.5,
  digits = 4
)
}
\arguments{
\item{y}{character. denotes the name of the column respectively to phenotype values}

\item{data}{data.frame. Should contain the following colunms: environemnt, genotype, phenotype.}

\item{random}{list A two-level list Specify the regression kernels (co-variance matrix). The former is the \code{Kernel},
where is included the regression kernels. The later is the \code{Type}, specifying if the matrix is either \code{D} Dense or
\code{BD} Block Diagonal. A number of regression kernels or random effects to be fitted are specified in this list.}

\item{fixed}{matrix Design matrix (\eqn{n \times p}) for fixed effects (NULL by default)}

\item{env}{character. denotes the name of the column respectively to environments}

\item{verbose}{Should iteration history be printed on console? If TRUE or 1 then it is printed,
otherwise, if another number $n$ is choosen the history is printed every $n$ times. The default is \code{FALSE}}

\item{iterations}{numeric Number of iterations (1000 by default)}

\item{burnin}{numeric, number of iterations to be discarded as burn-in (200 by default).}

\item{thining}{numeric, number of thining used in Markov Chains (10 by default)}

\item{tol}{a numeric tolerance level. Eigenvalues lower than \code{tol} are discarded. Default is \code{1e-10}.}

\item{R2}{the proportion of variance expected to be explained by the regression.}

\item{digits}{numeric. Digits for round variance components.}
}
\value{
A list contaning predicted phenotypes (yHat), posterior means of residual (VarE) and
 genetic/enviromic variance component for each term in the linear model with the respective confidence intervals (VarComp). Also the
 values along with the chains are released (BGGE).
}
\description{
Runs Bayesian Linear-Mixed Models for Multiple Environments using kernels from get_kernel function
}
\details{
TODO
}
\author{
Germano Costa Neto (Adapted from Granato et al. 2018, BGGE package)
}
