% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AtmosphericPAram.R
\name{Param_Atmospheric}
\alias{Param_Atmospheric}
\title{Support functions to estimate atmospheric parameters}
\usage{
Param_Atmospheric(
  PREC = NULL,
  Tdew = NULL,
  Tmin = NULL,
  Tmax = NULL,
  Alt = NULL,
  RH = NULL,
  Rad = NULL,
  G = NULL,
  alpha = 1.26,
  df,
  merge = FALSE
)
}
\arguments{
\item{K_E}{list of envirotype-related kernels (n x n genotypes-environment).
If NULL, benchmarck genomic kernels are built.}

\item{K_G}{list of genomic enabled kernels (p x p genotypes)}

\item{Y}{data.frame contaning the following colunms: environemnt, genotype, trait value}

\item{model}{model structure for genomic predicion. It can be c('MM','MDs','E-MM','E-MDs'),
which MM (main effect model or Y=fixed + G) amd MDs (Y=fixed+G+GxE)}

\item{reaction}{boolean, inclusion of a reaction norm based GxE kernel (default = FALSE)}

\item{intercept.random}{boolean, inclusion of a genomic random intercept (default = FALSE). For more details, see BGGE package vignette.}
}
\description{
Core of functions to estimate atmospheric parameters related to evapotranspiration. Fore more details about the equations see de FAO-evapotranspiration publication http://www.fao.org/3/X0490E/x0490e07.htm#atmospheric%20parameters
}
\author{
Germano Costa Neto
}
