% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AtmosphericPAram.R
\name{param_atmospheric}
\alias{param_atmospheric}
\title{Support functions to estimate atmospheric parameters}
\usage{
param_atmospheric(
  env.data,
  PREC = NULL,
  Tdew = NULL,
  Tmin = NULL,
  Tmax = NULL,
  RH = NULL,
  Rad = NULL,
  G = NULL,
  Alt = 600,
  alpha = 1.26,
  merge = FALSE
)
}
\arguments{
\item{env.data}{data.frame. A \code{get_weather()}-like output with additional solar radiation information. If a \code{get_weather()} is provided, no further argument is required but \code{G} and \code{alpha}.}

\item{PREC}{character. Indicates the column of precipitation.}

\item{Tdew}{character. Indicates the column of dew/frost.}

\item{Tmin}{character. Indicates the column of minimum temperature.}

\item{Tmax}{character. Indicates the column of maximum temperature.}

\item{RH}{character. Indicates the column of reative humidity.}

\item{Rad}{character. Indicates the column of solar radiation. This parameter can be calculated from \code{Param_Radiation()}.}

\item{G}{numeric. Flux of heat conducted into the ground. Default is 0.}

\item{Alt}{numeric. Elevation above sea level (meters)}

\item{alpha}{Alpha of Priestley & Taylor's (1972) equation. Default is 1.26, which fit data from most sources.}

\item{merge}{boolean. If \code{TRUE}, calculated variables are merged to the original \code{get_weather()} dataframe.}
}
\value{
A dataframe with parameters related to evapotranspiration. See details for further information.
}
\description{
Core of functions to estimate atmospheric parameters related to evapotranspiration. Fore more details about the equations see de FAO-evapotranspiration publication http://www.fao.org/3/X0490E/x0490e07.htm#atmospheric%20parameters.
}
\details{
This function requires a dataframe with all parameters listed above. If any is missing, an error will be returned.
The calculated variables are:
\itemize{
 \item SPV: Slope of saturation vapour pressure curve (kPa.Celsius)
 \item VPD: Vapour pressure deficit (kPa)
 \item ETP: Potential Evapotranspiration (mm.day)
 \item PEPT: Deficit by Precipitation (mm.day)
 }
}
\examples{
### Fetching weather information from NASA-POWER
env.data = get_weather(lat = -13.05, lon = -56.05, country = 'BRA')

### Calculating solar radiation
env.data = Param_Radiation(env.data, merge = TRUE)

### Calculating solar radiation
param_atmospheric(env.data)
}
\author{
Germano Costa Neto
}
