% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SradPARAM.R
\name{Param_Radiation}
\alias{Param_Radiation}
\title{Support functions to estimate basic radiation parameters}
\usage{
Param_Radiation(df, DOY = NULL, LAT = NULL, merge = FALSE)
}
\arguments{
\item{K_E}{list of envirotype-related kernels (n x n genotypes-environment).
If NULL, benchmarck genomic kernels are built.}

\item{K_G}{list of genomic enabled kernels (p x p genotypes)}

\item{Y}{data.frame contaning the following colunms: environemnt, genotype, trait value}

\item{model}{model structure for genomic predicion. It can be c('MM','MDs','E-MM','E-MDs'),
which MM (main effect model or Y=fixed + G) amd MDs (Y=fixed+G+GxE)}

\item{reaction}{boolean, inclusion of a reaction norm based GxE kernel (default = FALSE)}

\item{intercept.random}{boolean, inclusion of a genomic random intercept (default = FALSE). For more details, see BGGE package vignette.}
}
\description{
Core of functions to estimate atmospheric parameters related to solar radiation. Fore more details about the equations see de FAO-evapotranspiration publication http://www.fao.org/3/X0490E/x0490e07.htm#radiation
}
\author{
Germano Costa Neto
}
