% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processWTH.R
\name{processWTH}
\alias{processWTH}
\title{Basic Processing Tools for Enriching get_weather outputs}
\usage{
processWTH(
  x,
  lon = NULL,
  lat = NULL,
  env.id = NULL,
  DOY = NULL,
  download.ALT = TRUE,
  country = NULL
)
}
\arguments{
\item{x}{data.frame. A get_weather output}

\item{lon}{numeric. Longitude values in WGS84}

\item{lat}{numeric. Latitude values in WGS84}

\item{env.id}{character. Identification of the site or environment.}

\item{DOY}{numeric. Julian day. DOY column as benchmark (from get_weather)}

\item{download.ALT}{boolean. Default as TRUE, to download Altitude data from SRTM database of CGIAR.}

\item{country}{character. ID for country. Default is 'BRA'. For more detais see getData() from raster package.}
}
\description{
Summarize get_weather() outputs based on environments and defined time intervals (e.g.,phenology)
}
\author{
Germano Costa Neto
}
