% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processWTH.R
\name{processWTH}
\alias{processWTH}
\title{Basic Processing Tools for Enriching get_weather outputs}
\usage{
processWTH(weather.data)
}
\arguments{
\item{weather.data}{data.frame. A get_weather output}
}
\description{
A wraper of Param_Radiation(), Param_Atmospheric(), and Param_Temperature() for evaluating get_weather() datasets.
}
\author{
Germano Costa Neto
}
