% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Wmatrix.R
\name{W.matrix}
\alias{W.matrix}
\title{Build the Matrix of Environmental Covariables for Reaction Norm}
\usage{
W.matrix(
  weather.data,
  is.processed = FALSE,
  id.names = NULL,
  env.id = NULL,
  var.id = NULL,
  probs = NULL,
  by.interval = NULL,
  time.window = NULL,
  names.window = NULL,
  center = TRUE,
  scale = TRUE,
  sd.tol = 10,
  statistic = NULL,
  tol = 0.001,
  QC = FALSE
)
}
\arguments{
\item{weather.data}{data.frame of environmental variables gotten from \code{get_weather()}.}

\item{is.processed}{boolean. Indicates whether the dataframe was previously processed with \code{summaryWTH()} and contains means, medians, etc.}

\item{id.names}{character. Indicates the name of the columns to be used as id for the environmental variables to be analysed.}

\item{env.id}{character. Indicates the name of the columns to be used as id for environments.}

\item{var.id}{vector (character). Indicates which variables will be used in the analysis.}

\item{probs}{vector(numeric). Indicates the probability quantiles, as probs = {0,1}, if cardinals is NULL. If is quantiles=NULL, quantiles = c(0.01,.25,.50,.99).}

\item{by.interval}{boolean. Indicates if temporal intervals must be computed insied of each environment. Default = FALSE.}

\item{time.window}{vector (numeric). If \code{by.interval = TRUE}, this argument indicates the temporal breaks for delimited intervals.}

\item{names.window}{vector(character). If by.interval = TRUE, this argument indicates the names of the desirable intervals.}

\item{center}{boolean. Indicates whether the matrix should be centered.}

\item{scale}{boolean. If scale=TRUE, the variables assume a mean-centered scaled distribution, with \eqn{x~N(0,1)}.}

\item{sd.tol}{numeric. Default value equal to 10.}

\item{statistic}{vector (character). Indicates what statistic must be runned, statistic = c('all','sum','mean','quantile'). Default: 'mean'.}

\item{tol}{TODO.}

\item{QC}{boolean. Indicates with Quality Control is applied. QC is based on the standard deviation tolerance (sd.tol), removing variables (x) with sd(x) > sd.tol}
}
\value{
Returns a weather covariable realized matrix with dimensions of q x k, for q environments and k covariables.
}
\description{
Estimates a weather covariable realized matrix with dimensions of q x k, for q environments and k covariables.
}
\details{
TODO
}
\examples{
### Fetching weather information from NASA-POWER
weather.data = get_weather(lat = -13.05, lon = -56.05, country = 'BRA')

### Mean-centered and scaled matrix
W <- W.matrix(weather.data = weather.data, by.interval = FALSE)

## same as SummaryWTH, we can add time.windows
W <- W.matrix(weather.data = weather.data, by.interval = TRUE,
              time.window = c(0, 14, 35, 60, 90, 120))

## and select the statistic to be used
W <- W.matrix(weather.data = weather.data, by.interval = TRUE, statistic = 'mean',
              time.window = c(0, 14, 35, 60, 90, 120))

W <- W.matrix(weather.data = weather.data, by.interval = TRUE, statistic = 'quantile',
              time.window = c(0, 14, 35, 60, 90, 120))

## We can perform a Quality Control (QC) based on the maximum sd tolerated
W <- W.matrix(weather.data = weather.data, by.interval = FALSE, QC = TRUE)

## We can perform a Quality Control (QC) based on the maximum sd tolerated
W <- W.matrix(weather.data = weather.data, by.interval = FALSE, QC = TRUE, sd.tol = 3)

## or create W for specific variables
id.var = c('T2M_MAX','T2M_MIN','T2M')
W <- W.matrix(weather.data = weather.data, var.id = id.var)

## or combine with summaryWTH by using is.processed = TRUE
data <- summaryWTH(weather.data, env.id = 'env', statistic = 'quantile')
W <- W.matrix(weather.data = data, is.processed = TRUE)

}
\author{
Germano Costa Neto
}
